menambah validasi delete file dan menambah fitur delete bimbingan skripsi (#23)

+++ app/Traits/MediaTrait.php
@@ -19,6 +19,8 @@ trait MediaTrait
     }
 
     public function delete($path) {
-        Storage::disk('public')->delete($path);
+        if (Storage::disk('public')->exists($path)) {
+            Storage::disk('public')->delete($path);
+        }
     }
 }
